+++ src/client/js/index.js
@@ -2,4 +2,9 @@
  * pwix:tabbed/src/client/js/index.js
  */
 
+import { Instance } from '../classes/instance.class.js';
+
 import '../../common/js/index.js';
+
+Tabbed.Instance = Instance;
+Tabbed.instanceNames = {};